public/MusicBot/requirements.txt
python-telegram-bot>=20.7
yt-dlp
mutagen
requests
rich
ytmusicapi
beautifulsoup4
